Further work on the site.
The next task is to understand how cookies and sessions work; I feel
like I don't quite get it.  I also need to work with the strings that
are generated by START-SESSION.  Using SESSION-COOKIE-VALUE, I can get
the 32-bit hex string.  One problem is that I need to use this data to
create the appropriate directory.  It has a colon in it, so I need to
strip that away.