In a subdirectory, add a patched version of Edi Weitz's CL-WHO, modified by Victor Kryukov.
Kryukov's patches allow macros (suitably defined, using
DEF-INTERNAL-MACRO, rather than with simpy DEFMACRO) to be used within
HTM forms.  I'm making this a git submodule.  The ASDF system
definition of TEXSERV is adjusted to depend on this system.  That way,
the other versions of CL-WHO that I have sitting around (under ~/.sbcl
and the CLBUILD directories, for example) to be
unaffected.  (Previously, I was patching these "canonical" versions of
CL-WHO with Victor's changes.  Then I fooled around a bit with CLBUILD
and got bit by some compiler troubles which were caused by the fact
that I was using the canonical version of CL-WHO rather than the
patched version that I had come to rely on for some time.)